--- LazyNecromancer/Assets/Scripts/Rats/BoidPrefab.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6b4ad74547957154c843b013932debab
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
