--- LazyNecromancer/Assets/Scripts/Rats/Cell.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b671e805d5429a47b7fd56d66f9407a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
