--- LazyNecromancer/Assets/Scripts/Rats/GregBoid.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2fb05f886d3ed0e409ae21689c385958
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
