--- LazyNecromancer/Assets/Scripts/Rats/GregBoidController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2276a468679f52d4684b82ebd275e9a0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
