--- LazyNecromancer/Assets/Scripts/Rats/Grid.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5c8bafc8cd619c43bdab1de7d163372
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
