--- LazyNecromancer/Assets/Scripts/Rats/PBoid.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 703a1d9814bafb545abb39de03138d20
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
