--- LazyNecromancer/Assets/Scripts/Rats/RatPrefab.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5e15369f1e18f8a4b9cf51fa3acbe73e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
