+++ LazyNecromancer/Assets/Scripts/Rats/Target.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Target : MonoBehaviour {
+
+    void Update() {
+        transform.position = Camera.main.ScreenToWorldPoint(Input.mousePosition);
+    }
+}
