+++ LazyNecromancer/Assets/Scripts/Rats/Target.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 857f8e3306361024eb8c51d5ff63eeb3
+guid: 0b00452114df0934c81a509abbcc0b9c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
